Issue #43 - Get self name properly
