+++ robo.yml.dist
@@ -1,6 +1,5 @@
-
 php:
     executables:
-        current:
+        php_min:
             enabled: true
             command: 'php'
